Upload go binaries as functions
With AWSLambda supporting Go, I thought it'd be interesting to
do some exploration to support them.

Go support requires binary files, so we cannot depend on the file
extension to know if what we're uploading is JS or Go. Instead, we can
have a well known path convention to put Go binaries, like
`well-known-go`, and assume that all executables there are Go binaries.

We'd also need a way to tell our API that something should use the go
runtime and not the js runtime. I added a `Runtime` attribute to the file
bundle, so we can configure aws lambda with the right flags.

Signed-off-by: David Calavera <david.calavera@gmail.com>