feat: remove site_capabilities from the deploy object
This was used just for large media and we're deprecating it. We'll be
removing this slowly and this is a good first step and allows us to
return a smaller object for deploys.

Updated the porcelain handler to not rely on this field but only on the
option passed on DoDeploy. This way we allow for builds to still use it
if needed until fully retired.